Cleaner output and updated drop timing
I cleaned up the output on some of the querying functions in the clever
wrapper library. This makes for a more readable output in the console
window. I also changed the way the library waits for the end of songs
and drops. Now it takes into account when playback is paused and played,
rewound or fast forwarded.